+++ src/components/atomic-component/enchanted-badge.ts
@@ -20,16 +20,39 @@ import { customElement, property } from 'lit/decorators.js';
 import { EnchantedAcBaseElement } from './enchanted-ac-base-element';
 
 // Helper imports
-import { BADGE_PARTS } from '../../types/cssClassEnums';
-import { isLTR } from '../localization';
+import { getCurrentDirection } from "../localization";
+import { LOCALE_DIRECTIONS } from "../constants";
+import { EnchantedBadgeColor, EnchantedBadgeBorder, EnchantedBadgeType, EnchantedBadgeParts } from '../../types/cssClassEnums';
  
 @customElement('enchanted-badge')
 export class EnchantedBadge extends EnchantedAcBaseElement {
 
-  @property({ type: String }) badge = '';
-  
+
+  @property({ reflect: true }) color: EnchantedBadgeColor = EnchantedBadgeColor.PRIMARY;
+
+  @property({ reflect: true }) border: EnchantedBadgeBorder = EnchantedBadgeBorder.DEFAULT;
+
+  @property({ reflect: true }) badge: EnchantedBadgeType = EnchantedBadgeType.TEXT;
+
+  @property() text: string = ''; // Added property to allow user to pass any string for badge:text
+
+
   render() {
-    return html`<div part=${isLTR() ? BADGE_PARTS.BADGE_TEXT : BADGE_PARTS.BADGE_TEXT_RTL} data-testid="enchanted-badge">${this.badge}</div>`;
+    const isDirectionLTR = getCurrentDirection() === LOCALE_DIRECTIONS.LTR;
+    const partName = this.badge === EnchantedBadgeType.DOT
+      ? EnchantedBadgeParts.BADGE_DOT
+      : isDirectionLTR
+        ? EnchantedBadgeParts.BADGE_TEXT
+        : EnchantedBadgeParts.BADGE_TEXT_RTL;
+
+    return html`
+      <div
+        part="${partName}"
+        data-testid="enchanted-badge"
+      >
+        ${this.badge === EnchantedBadgeType.TEXT ? this.text : ''}
+      </div>
+    `;
   }
 }
  
@@ -38,4 +61,3 @@ declare global {
     'enchanted-badge': EnchantedBadge
   }
 }
-